requirements.txt
# Core
# streamlit>=1.28.0
# groq>=0.9.0
# reportlab>=4.0.0

# # RAG
# pypdf>=4.0.0
# sentence-transformers>=2.7.0
# faiss-cpu>=1.8.0
# numpy>=1.26.0

# # LangGraph
# langgraph>=0.2.0



streamlit>=1.28.0
groq>=0.9.0
reportlab>=4.0.0
pypdf>=4.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
langgraph>=0.2.0
